chore: made folders for program files
