+++ lesson_25/kimberlees-template/src/pages/Home/Home.tsx
@@ -1,13 +1,32 @@
 import './Home.scss';
-import ProgramList from './programlist.tsx';
+import ProgramList from './program-list/programlist.tsx';
 import React from 'react';
 
-
 const programsData = [
-  { id: 1, title: 'Swine Short Loin', description: 'Swine short loin burgdoggen ball tip, shank ham hock bacon. Picanha strip steak pork, swine boudin ham meatball meatloaf leberkas sausage. Turkey beef andouille kielbasa rump pastrami biltong chislic alcatra buffalo prosciutto jowl. Pastrami chicken sirloin swine capicola landjaeger jowl boudin pork chop shankle bresaola turducken leberkas ham.' },
-  { id: 2, title: 'Bacon Ipsum', description: 'Bacon ipsum dolor amet leberkas chuck biltong pork loin sirloin bresaola rump frankfurter. Shoulder doner strip steak chuck. Short ribs venison salami chuck sirloin jowl chislic cupim swine cow. Corned beef chuck frankfurter tenderloin venison biltong andouille leberkas kielbasa sausage t-bone turducken fatback. Pork picanha buffalo bacon tail salami meatball, jowl chislic.' },
-  { id: 3, title: 'Picanha Swine Jowl', description: 'Picanha swine jowl meatball boudin pastrami bresaola fatback shankle pork belly cow. Frankfurter ground round shank corned beef chuck. Jerky frankfurter fatback capicola hamburger, pork prosciutto bresaola ham porchetta rump t-bone pancetta tenderloin. Fatback ham hock prosciutto, tenderloin shoulder salami tri-tip leberkas bacon turducken chislic venison sausage frankfurter.' },
-  { id: 4, title: 'Kevin Chicken T-Bone', description: 'Kevin chicken t-bone spare ribs shankle bacon drumstick kielbasa cow jowl doner salami chuck andouille. Rump spare ribs bresaola frankfurter shankle. Bresaola ribeye turducken, cow leberkas boudin biltong sirloin filet mignon ham pork belly shank. Tenderloin sirloin pancetta pork loin shankle venison turducken boudin. Brisket tenderloin salami ham bresaola burgdoggen.' }
+  {
+    id: 1,
+    title: 'Swine Short Loin',
+    description:
+      'Swine short loin burgdoggen ball tip, shank ham hock bacon. Picanha strip steak pork, swine boudin ham meatball meatloaf leberkas sausage. Turkey beef andouille kielbasa rump pastrami biltong chislic alcatra buffalo prosciutto jowl. Pastrami chicken sirloin swine capicola landjaeger jowl boudin pork chop shankle bresaola turducken leberkas ham.',
+  },
+  {
+    id: 2,
+    title: 'Bacon Ipsum',
+    description:
+      'Bacon ipsum dolor amet leberkas chuck biltong pork loin sirloin bresaola rump frankfurter. Shoulder doner strip steak chuck. Short ribs venison salami chuck sirloin jowl chislic cupim swine cow. Corned beef chuck frankfurter tenderloin venison biltong andouille leberkas kielbasa sausage t-bone turducken fatback. Pork picanha buffalo bacon tail salami meatball, jowl chislic.',
+  },
+  {
+    id: 3,
+    title: 'Picanha Swine Jowl',
+    description:
+      'Picanha swine jowl meatball boudin pastrami bresaola fatback shankle pork belly cow. Frankfurter ground round shank corned beef chuck. Jerky frankfurter fatback capicola hamburger, pork prosciutto bresaola ham porchetta rump t-bone pancetta tenderloin. Fatback ham hock prosciutto, tenderloin shoulder salami tri-tip leberkas bacon turducken chislic venison sausage frankfurter.',
+  },
+  {
+    id: 4,
+    title: 'Kevin Chicken T-Bone',
+    description:
+      'Kevin chicken t-bone spare ribs shankle bacon drumstick kielbasa cow jowl doner salami chuck andouille. Rump spare ribs bresaola frankfurter shankle. Bresaola ribeye turducken, cow leberkas boudin biltong sirloin filet mignon ham pork belly shank. Tenderloin sirloin pancetta pork loin shankle venison turducken boudin. Brisket tenderloin salami ham bresaola burgdoggen.',
+  },
 ];
 
 export const Home: React.FC = () => {
@@ -30,55 +49,6 @@ export const Home: React.FC = () => {
       </section>
       <section className="programs-section">
         <ProgramList programs={programsData} />
-        {/*<h2>
-          Our <em className="highlight">Programs</em>
-        </h2>
-        <ul className="programs">
-          <li className="program">
-            <h3>Swine Short Loin</h3>
-            <p>
-              Swine short loin burgdoggen ball tip, shank ham hock bacon.
-              Picanha strip steak pork, swine boudin ham meatball meatloaf
-              leberkas sausage. Turkey beef andouille kielbasa rump pastrami
-              biltong chislic alcatra buffalo prosciutto jowl. Pastrami chicken
-              sirloin swine capicola landjaeger jowl boudin pork chop shankle
-              bresaola turducken leberkas ham.
-            </p>
-          </li>
-          <li className="program">
-            <h3>Bacon Ipsum</h3>
-            <p>
-              Bacon ipsum dolor amet leberkas chuck biltong pork loin sirloin
-              bresaola rump frankfurter. Shoulder doner strip steak chuck. Short
-              ribs venison salami chuck sirloin jowl chislic cupim swine cow.
-              Corned beef chuck frankfurter tenderloin venison biltong andouille
-              leberkas kielbasa sausage t-bone turducken fatback. Pork picanha
-              buffalo bacon tail salami meatball, jowl chislic.
-            </p>
-          </li>
-          <li className="program">
-            <h3>Picanha Swine Jowl</h3>
-            <p>
-              Picanha swine jowl meatball boudin pastrami bresaola fatback
-              shankle pork belly cow. Frankfurter ground round shank corned beef
-              chuck. Jerky frankfurter fatback capicola hamburger, pork
-              prosciutto bresaola ham porchetta rump t-bone pancetta tenderloin.
-              Fatback ham hock prosciutto, tenderloin shoulder salami tri-tip
-              leberkas bacon turducken chislic venison sausage frankfurter.
-            </p>
-          </li>
-          <li className="program">
-            <h3>Kevin Chicken T-Bone</h3>
-            <p>
-              Kevin chicken t-bone spare ribs shankle bacon drumstick kielbasa
-              cow jowl doner salami chuck andouille. Rump spare ribs bresaola
-              frankfurter shankle. Bresaola ribeye turducken, cow leberkas
-              boudin biltong sirloin filet mignon ham pork belly shank.
-              Tenderloin sirloin pancetta pork loin shankle venison turducken
-              boudin. Brisket tenderloin salami ham bresaola burgdoggen.
-            </p>
-          </li>
-        </ul>*/}
       </section>
     </article>
   );
